Fix to create and attach a HM when HM reference or template is removed from GSLB hostrule and no HM reference or template is given in the GDP object.
