+++ gslb/ingestion/gslb.go
@@ -22,10 +22,13 @@ import (
 	"os"
 	"sort"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 
+	"github.com/vmware/alb-sdk/go/models"
 	"github.com/vmware/global-load-balancing-services-for-kubernetes/gslb/apiserver"
+	avictrl "github.com/vmware/global-load-balancing-services-for-kubernetes/gslb/cache"
 	"github.com/vmware/global-load-balancing-services-for-kubernetes/gslb/gslbutils"
 	"github.com/vmware/global-load-balancing-services-for-kubernetes/gslb/nodes"
 
@@ -952,4 +955,67 @@ func loadClusterAccess(membersKubeConfig string, memberClusters []gslbalphav1.Me
 		gslbutils.Logf("cluster: %s, msg: %s", memberCluster.ClusterContext, "loaded cluster access")
 	}
 	return clusterDetails
+}
+
+func isHealthMonitorTemplatePresentInCache(hmName string) bool {
+	aviHmCache := avictrl.GetAviHmCache()
+
+	obj, present := aviHmCache.AviHmCacheGet(avictrl.TenantName{Tenant: utils.ADMIN_NS, Name: hmName})
+	if !present {
+		return false
+	}
+	hmObj, ok := obj.(*avictrl.AviHmObj)
+	return ok && hmObj.CustomHmSettings != nil
+}
+
+// Checks whether the template is present in the controller. If present, it adds the contents of the template to the cache which will
+// used for the creation of health monitors. If the template is not present in the controller, the GDP/GSLBHostRule
+// will be rejected.
+func validateAndAddHmTemplateToCache(hmTemplate string, gdp bool, fullSync bool) error {
+	if fullSync && isHealthMonitorTemplatePresentInCache(hmTemplate) {
+		gslbutils.Debugf("health monitor template %s present in hm cache", hmTemplate)
+		return nil
+	}
+	hm, err := getHMFromName(hmTemplate, gdp)
+	if err != nil {
+		gslbutils.Errf("Health Monitor Template %s not found", hmTemplate)
+		return fmt.Errorf("health monitor template %s not found", hmTemplate)
+	}
+	if hm.IsFederated != nil &&
+		!(*hm.IsFederated) {
+		gslbutils.Errf("Health Monitor Template %s not federated", hmTemplate)
+		return fmt.Errorf("health monitor template %s not federated", hmTemplate)
+	}
+	// Get the response code, request header based on the protocol type of HM.
+	var hmHTTP *models.HealthMonitorHTTP
+	switch *hm.Type {
+	case gslbutils.SystemGslbHealthMonitorHTTP:
+		hmHTTP = hm.HTTPMonitor
+	case gslbutils.SystemGslbHealthMonitorHTTPS:
+		hmHTTP = hm.HTTPSMonitor
+	default:
+		return fmt.Errorf("health monitor template is not supported for non-path based health monitors")
+	}
+
+	// client request header must be in the format <GET/HEAD> /<path> <HTTP/version>
+	if len(strings.Split(*hmHTTP.HTTPRequest, gslbutils.RequestHeaderStringSeparator)) != gslbutils.NoOfRequestHeaderParams {
+		gslbutils.Errf("Client request header in Health Monitor Template %s is not correct", hmTemplate)
+		return fmt.Errorf("client request header in health monitor template %s is invalid", hmTemplate)
+	}
+
+	key := avictrl.TenantName{Tenant: utils.ADMIN_NS, Name: hmTemplate}
+	hmCacheObj := avictrl.AviHmObj{
+		Name:   hmTemplate,
+		Tenant: utils.ADMIN_NS,
+		UUID:   *hm.UUID,
+		CustomHmSettings: &avictrl.CustomHmSettings{
+			RequestHeader: *hmHTTP.HTTPRequest,
+			ResponseCode:  hmHTTP.HTTPResponseCode,
+		},
+		Description: gslbutils.CreatedByUser,
+		CreatedBy:   gslbutils.CreatedByUser,
+	}
+	aviHmCache := avictrl.GetAviHmCache()
+	aviHmCache.AviHmCacheAdd(key, &hmCacheObj)
+	return nil
 }
