+++ gslb/nodes/dq_ingestion.go
@@ -187,7 +187,7 @@ func AddUpdateGSLBHostRuleOperation(key, objType, objName string, wq *utils.Work
 	prevHmChecksum := GetHmChecksum(objType, gsGraph)
 	prevChecksum := gsGraph.GetChecksum()
 	// update the GS graph
-	aviGS.(*AviGSObjectGraph).UpdateAviGSGraphWithGSFqdn(objName, false, gsGraph.MemberObjs[0].TLS)
+	aviGS.(*AviGSObjectGraph).UpdateAviGSGraphWithGSFqdn(key, objName, false, gsGraph.MemberObjs[0].TLS)
 	newChecksum := gsGraph.GetChecksum()
 	newHmChecksum := GetHmChecksum(objType, gsGraph)
 	gslbutils.Debugf("prevChecksum: %d, newChecksum: %d, prevHmChecksum: %d, newHmChecksum: %d, key: %s", prevChecksum,
